frontend: router: Break self-referencing barrel import in getRoutePath
getRoutePath.tsx imported NotFoundRoute from the router's own barrel
(lib/router/index.tsx) just to read its literal '*' path. That barrel
imports nearly every page component in the app, including several
that pull in the lib/k8s barrel.

Any resource class importing KubeObject (directly or through the k8s
barrel) reaches KubeObject.ts, which imports createRouteURL, which
transitively reaches getRoutePath.tsx. When that happens before
KubeObject.ts itself has finished initializing (e.g. a test importing
a resource class in isolation, without the app first bootstrapping),
the self-reference drags the whole router barrel in with it, which in
turn re-enters the k8s barrel and tries to build a resource class that
extends KubeObject before KubeObject exists, crashing with
'Class extends value undefined is not a constructor or null'.

Move NotFoundRoute into its own small file with no dependency on the
router barrel, and have getRoutePath.tsx import it from there instead
of '.'. router/index.tsx re-exports it so existing consumers are
unaffected.

Two tests relied on this cycle to accidentally trigger the router
barrel's setDefaultRoutes() call as a side effect of importing
createRouteURL; now that the cycle is gone they need to import the
router barrel explicitly to get routes registered.

Added a regression test that imports a resource class from a
components/ test file, matching the crash's original repro.
