+++ frontend/src/lib/router/notFoundRoute.tsx
@@ -14,18 +14,16 @@
  * limitations under the License.
  */
 
-import { getClusterPrefixedPath } from '../cluster';
-import { getRouteUseClusterURL } from './getRouteUseClusterURL';
-import { NotFoundRoute } from './notFoundRoute';
-import type { Route } from './Route';
+import NotFoundComponent from '../../components/404';
 
-export function getRoutePath(route: Route) {
-  if (route.path === NotFoundRoute.path) {
-    return route.path;
-  }
-  if (!getRouteUseClusterURL(route)) {
-    return route.path;
-  }
-
-  return getClusterPrefixedPath(route.path);
-}
+// The NotFound route needs to be considered always in the last place when used
+// with the router switch, as any routes added after this one will never be considered.
+// So we do not include it in the default routes in order to always "manually" consider it.
+export const NotFoundRoute = {
+  path: '*',
+  exact: true,
+  name: `Whoops! This page doesn't exist`,
+  component: () => <NotFoundComponent />,
+  sidebar: null,
+  noAuthRequired: true,
+};
